chore: move dependencies out of routers package
