+++ api/pybot/main.py
@@ -10,9 +10,9 @@ from fastapi.responses import JSONResponse
 from fastapi.staticfiles import StaticFiles
 from loguru import logger
 
+from pybot.dependencies import UserIdHeader
 from pybot.routers.chat import router as chat_router
 from pybot.routers.conversations import router as conversation_router
-from pybot.routers.dependencies import UserIdHeader
 from pybot.routers.files import router as files_router
 from pybot.routers.messages import router as message_router
