+++ api/pybot/routers/conversations.py
@@ -9,9 +9,9 @@ from loguru import logger
 
 from pybot.config import settings
 from pybot.context import session_id
+from pybot.dependencies import ChatMemory, Llm, MessageHistory, UserIdHeader
 from pybot.jupyter import ContextAwareKernelManager, GatewayClient
 from pybot.models import Conversation as ORMConversation
-from pybot.routers.dependencies import ChatMemory, Llm, MessageHistory, UserIdHeader
 from pybot.schemas import (
     ChatMessage,
     Conversation,
